HUD and Boulder Designed + Partial Implementation
Organization::
-Created files and organized everything in github
-Fixed spacing and organized code

Added::
- The HUD now displays on the screen and is partially implemented. We
will add more stuff to the HUD as we add more things to the project.
- The boulder class was implemented and the boulder now displays on the
screen. It also drops when the dirt underneath it is destroyed. It does
not damage diggerman at this point.
- Health variable given to diggerman to keep track of his health status
- Many functions have been shortened as we have come up with faster
computational solutions.
-The temporary and permanent nugget classes were created and partially
designed. They will be implemented soon.
-Mathematical expressions for determining how many gold nuggets,
boulders, and oil barrels will spawn in each level were created. They
are called numOfGoldNuggets(), numOfBoulders(), and numOfOilBarrels().