Updated how dirt is handled
Updated how dirt is handled (how it is filled and removed in the grid/arena). In  SW.CPP created a fillDirt function that
handles the filling of dirt, which is only ever called in init() and
updated the function to make it more efficient.

Updated the remove dirt to run more efficient, and it also deletes the
dirt objects from the dirt array if they are actually present instead of just making it invisible. This
function can be used by any 4x4 assist to delete dirt around it, if need
be. Done by Jacob Frick, re-committed by me.