Wrote in protester random moving + small fixes
-Fixed bug in the check dirt around functions in protester class, as
they had wrong numbers adding to getX() and getY()to properly checkc for
dirt in that direction
-Added protester random moving. It checks the direction that it is
facing, makes sure its next move won't be out of bounds and that there
is not dirt in that direction, if all cases fit, moves in that
direction, then waits to proper amount of ticks. If it fails those 4
cases, it just sets a new direction and trys again
-added function to compute a random move count for protester that
returns 8 <=to >= 60
-put all of the move contents of protester into a moveProtester()
function for easy of readability of doSomething()
-also added a moveProtesterHelper(int x, int y) function, this is meant
only for the moveProtester() function. meant for moving 1 square in one
specific direction and decs the move counter.