Protester Update + minor changes + code cleanup
-Protester now follows digger man if they are on the same x or y access,
if there is no dirt bewteen them
-Also created a helper function for the following of dm
-Protester now shouts at dm and dm loses health. If digger man loses all
health, he loses a life. Though there is a glitch where the old dm and
protester stay there and new ones are created, not sure why yet
-Moved the check dirt around the actor functions to sw because better
design. Also gave them an integer arg to check that many more spaces in
that direction. So passing 0 to the isDirtUnderMe would result in
directly under, passing 1, would 1 more space down, etc.
-Updated the DMInVicinity function in student world to take an x and y
cord, insead of an actor because better design. Also updated the
functions that used it that were written by the goup so they pass the x
and y instead of themselves via this