Protester now follows quickest path
Got the protester follow the quickest to (60,60) path when he is
annoyed! I set up an array that looks at the entire grid at that current
moment. It fills a second 64x64 grid with -2 if it is a space that the
protester cannot occupy.  Which means it cannot occupy that cord if the
entire 4x4 protester collides with any dirt at all. It sets it with -1
if can occupy that space. Used breadth first search to find quickest
path