Fixed protester glitch + update
Fixed glitch with quickpath finder where it caused infinite loop when to
much dirt was removed. Also made protester change to face the proper
direction when following the path.  Also when protester hits (60,60) he
will be set to dead and stop moving. Still need to add protester to
vector of actors in order to be removed