Water pool randomly spawns, Boulder kills actors
-made a new function thats kinda redundent called isDirtless. I could've
used isThereDirtVisible here but making this new one made it easier for
me to make sure the coord and the 4x4 radius it covers are free of
dirt/I was too lazy to change the other one.
-Diggerman and Protestors will get die/annoyed when hit by a falling
boulder.
-Water pool spawns in a random place where there is no dirt.
-Some bugs I noticed:
-There is a weird bug with the protestor where he sometimes stays stuck
choosing between which direction to go after he gets annoyed by a
falling boulder. Will provide vid.
-The game will go into an infinite loop when trying to decide where to
put a new water pool. This only happens after losing the first life and
it sometimes doesn't happen at all.
-Sometimes after the restart of a level after losing a life, a random
gold nugget or boulder or protestor will spawn where they're not
supposed to (ex. a protestor in the middle of dirt in the middle of the
field at the start of a level).